extract common types to parent folder then import js_only
